requirements.txt
langchain
langchain-community
langgraph
langgraph-supervisor
langchain-groq
langchain-openai
duckduckgo-search
python-dotenv
ipykernel